% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mm.R
\name{import_mm}
\alias{import_mm}
\title{import DEB2 match mutate model from a google spreadsheet}
\usage{
import_mm(gsheet = NULL)
}
\arguments{
\item{gsheet}{name of sheet using spreadsheet/worksheet syntax
in the users googledrive}
}
\value{
a pov object of class \code{de_pov}.
}
\description{
match mutate model
}
\examples{
\dontrun{
# We create an
# empty pov object (and add in nodes and edges later,
# with other functions)
# TODO: Change this to deb2_mm_example later
mmm <- import_mm(gsheet = "amaffei_mm_whoifsr_03/whoifsr_itdcsv_2019")
mmm
}
}
