% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{mm}
\alias{mm}
\title{Mutate input tibble based on match rules found in a match-mutate tibble}
\usage{
mm(input_tbl, mm_table)
}
\arguments{
\item{input_tbl}{table of observations to be matched}

\item{mm_table}{table of match/replace rules}
}
\value{
tibble
}
\description{
Mutate input tibble based on match rules found in a match-mutate tibble
}
\examples{
# This function is based on a python script written by Andrew Maffei and
# Joe Futrelle while they worked at WHOI
joes_function <- function(string){
  library(stringr)
  lookup_tbl <- tibble::tribble(
    ~name, ~val,
    "AMAFF", "Andrew Maffei",
    "JFUTR", "Joe Futrelle")
  ans <- lookup_tbl \%>\%
    dplyr::filter (name == string) \%>\%
    dplyr::select (val) \%>\%
    toString()
  return(ans)
}

observed_data_tbl  <- tibble::tribble(
  ~date, ~expcode, ~desc, ~ref, ~amt,
  "2019-01-01", "5170", "AMAFF-SANFRAN-21320","","329.00",
  "2019-01-02", "5210", "PO# 79342 to Staples", "AMAFF", "92.64",
  "2019-01-03", "5170", "Car Rental Refund","","120.32")

rules_tbl <- tibble::tribble(
  ~date, ~expcode, ~desc, ~ref, ~amt, ~category, ~object, ~place, ~po, ~destination, ~traveler,
  "","(?<expcode>5170)","","","","Travel:Domestic:Unknown","","Woods Hole", "","","",
  "","(?<expcode>5170)","(?<trav>[^-]+)-(<dest>[^-]+)-(?<po>\\\\d{5})","","","Travel:Domestic:{trav}","TravelVendor","","{po}", "{dest}","{joes_function(trav)}",
  "","(?P<expcode>5210)","","","","Equipment:Unknown","","","","","",
  "","(?P<expcode>5210)","PO# (?<po>\\\\d{5}) to Staples","(?<ref>.*)","","Equipment:Staples:{ref}","Staples","","{po}","",""
)


mm(observed_data_tbl, rules_tbl)
}
