% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{mm}
\alias{mm}
\title{Mutate input tibble based on match rules found in a match-mutate rules tibble}
\usage{
mm(observations_tbl, rules_tbl)
}
\arguments{
\item{observations_tbl}{table of observations to be matched}

\item{rules_tbl}{table of match/replace rules}
}
\value{
tibble
}
\description{
Mutate input tibble based on match rules found in a match-mutate rules tibble
}
\examples{
# This function is based on a python script written by Andrew Maffei and
# Joe Futrelle while they worked at WHOI
joes_function <- function(string){
  library(stringr)
  lookup_tbl <- tibble::tribble(
    ~name, ~val,
    "AMAFF", "Andrew Maffei",
    "JFUTR", "Joe Futrelle")
  ans <- lookup_tbl \%>\%
    dplyr::filter (name == string) \%>\%
    dplyr::select (val) \%>\%
    toString()
  return(ans)
}

transactions  <- tibble::tribble(
  ~date, ~expcode, ~desc, ~ref, ~amt,
  "2019-01-01", "5170", "AMAFF-SANFRAN-21320","","329.00",
  "2019-01-02", "5210", "PO# 79342 to Staples", "AMAFF", "92.64",
  "2019-01-03", "5170", "Car Rental Refund","","120.32")

rules <- tibble::tribble(
  ~date, ~expcode, ~desc, ~ref, ~amt, ~category, ~object, ~xactloc, ~po, ~destination, ~person,
  "","(?<expcode>5170)","","","","Travel:Domestic:Unknown","Unknown","Unknown", "Unknown","Unknown","Unknown",
  "","(?<expcode>5170)","(?<person>[^-]+)-(?<dest>[^-]+)-(?<ponum>\\\\d{5})","","","Travel:Domestic:{person}","TravelVendor","WWW","{ponum}", "{dest}","{joes_function({person})}",
  "","(?<expcode>5210)","Unknown","Unknown","Unknown","Equipment:Unknown","Unknown","Unknown","Unknown","Unknown","Unknown",
  "","(?<expcode>5210)","PO# (?<ponum>\\\\d{5}) to Staples","(?<person>.*)","","Supplies:Staples:Rocketbooks","Staples","Falmouth","{ponum}","Unknown","{joes_function({person})}"
)
mm(transactions, rules)

}
